Skip hawkrest authentication if it's not detected and not mandatory.
To allow hawkrest to play well with other authentication schemes
it should return None if a proper hawk HTTP_AUTHORIZATION header
is not detected.
To avoid breaking backward compatibility I hid the detection behind
a HAWK_IS_MANDATORY setting which defaults to True.
For security reasons, in case of a missing Authorization header
an exception is raised anyway.