Stop testing Django 1.6/1.7 against Python 3
Both mainstream and extended support has now ended for Django 1.6/1.7:
https://www.djangoproject.com/download/#supported-versions

Now that we're shortly going to test on Python 3.5, Django 1.9 and
django-rest-framework 3.3, the testing matrix is going to get very large
unless we cut back on testing for EOL Django versions.

It's unlikely people stuck on older versions of Django are using
Python 3, so this seems like an easy win.